+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Application/Entities/Item.cs
@@ -1,7 +1,6 @@
-using System;
 using Amazon.DynamoDBv2.DataModel;
 
-namespace DotNetServerless.Domain.Entity
+namespace DotNetServerless.Application.Entities
 {
   public class Item
   {
