+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Application/Requests/CreateItemRequest.cs
@@ -1,11 +1,10 @@
-using DotNetServerless.Domain.Entity;
+using DotNetServerless.Application.Entities;
 using MediatR;
 
-namespace DotNetServerless.Domain.Requests
+namespace DotNetServerless.Application.Requests
 {
-  public class UpdateItemRequest : IRequest<Item>
+  public class CreateItemRequest : IRequest<Item>
   {
-    public string Id { get; set; }
     public string Description { get; set; }
     public string Code { get; set; }
     public bool IsChecked { get; set; }
@@ -14,7 +13,6 @@ namespace DotNetServerless.Domain.Requests
     {
       return new Item
       {
-        Id = Id,
         Description = Description,
         Code = Code,
         IsChecked = IsChecked
