+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Application/Responses/ItemResponse.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace DotNetServerless.Domain.Responses
+namespace DotNetServerless.Application.Responses
 {
   public class ItemResponse
   {
