+++ aws-dotnet-rest-api-with-dynamodb/src/DotNetServerless.Lambda/DotNetServerless.Lambda.csproj
@@ -29,7 +29,7 @@
   </ItemGroup>
   
   <ItemGroup>
-    <ProjectReference Include="..\DotNetServerless.Domain\DotNetServerless.Domain.csproj" />
+    <ProjectReference Include="..\DotNetServerless.Application\DotNetServerless.Application.csproj" />
   </ItemGroup>
 
 </Project>
